+++ src/App.js
@@ -1,19 +1,53 @@
-import React, { Component } from 'react';
-import logo from './logo.svg';
-import './App.css';
-
+import React, { Component } from "react";
+import "./App.css";
+import Counters from "./components/counters";
+import NavBar from "./components/navBar";
 class App extends Component {
+  state = {
+    counters: [
+      { id: 1, value: 4 },
+      { id: 2, value: 0 },
+      { id: 3, value: 0 },
+      { id: 4, value: 0 }
+    ]
+  };
+
+  handleDelete = counterID => {
+    const counters = this.state.counters.filter(c => c.id !== counterID);
+    this.setState({ counters });
+  };
+
+  handleReset = () => {
+    const counters = this.state.counters.map(c => {
+      c.value = 0;
+      return c;
+    });
+    this.setState({ counters });
+  };
+
+  handleIncrement = counter => {
+    const counters = [...this.state.counters];
+    const index = counters.indexOf(counter);
+    //if dont have this line. It will update the state first before setting the state.
+    counters[index] = { ...counter };
+    counters[index].value++;
+    //console.log(this.state.counters);
+    this.setState({ counters });
+  };
+
   render() {
     return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
+      <React.Fragment>
+        <NavBar
+          counterTotal={this.state.counters.filter(c => c.value > 0).length}
+        />
+        <Counters
+          counters={this.state.counters}
+          onIncrement={this.handleIncrement}
+          onDelete={this.handleDelete}
+          onReset={this.handleReset}
+        />
+      </React.Fragment>
     );
   }
 }
