+++ src/components/counters.jsx
@@ -2,52 +2,22 @@ import React, { Component } from "react";
 import Counter from "./counter";
 
 class Counters extends Component {
-  state = {
-    counters: [
-      { id: 1, value: 4 },
-      { id: 2, value: 0 },
-      { id: 3, value: 0 },
-      { id: 4, value: 0 }
-    ]
-  };
-
-  handleDelete = counterID => {
-    const counters = this.state.counters.filter(c => c.id !== counterID);
-    this.setState({ counters });
-  };
-
-  handleReset = () => {
-    const counters = this.state.counters.map(c => {
-      c.value = 0;
-      return c;
-    });
-    this.setState({ counters });
-  };
-  handleIncrement = counter => {
-    const counters = [...this.state.counters];
-    //counters[1].value++;
-    const index = counters.indexOf(counter);
-    counters[index] = { ...counter };
-    counters[index].value++;
-    console.log(this.state.counters);
-    this.setState({ counters });
-  };
-
   render() {
     return (
-      <div>
+      <div className="m-4">
         <button
-          onClick={this.handleReset}
+          onClick={this.props.onReset}
           className="btn btn-primary btn-sm m-2"
         >
           Reset
         </button>
-        {this.state.counters.map(counter => (
+        {this.props.counters.map(counter => (
           <Counter
             key={counter.id}
             counter={counter}
-            onDelete={this.handleDelete}
-            onIncrement={this.handleIncrement}
+            onDelete={this.props.onDelete}
+            onIncrement={this.props.onIncrement}
+            onReset={this.props.onReset}
           />
         ))}
       </div>
